Fix leap years.
A normal year has 365 days.
A Leap Year has 366 days (the extra day is the 29th of February).
yes	Leap Years are any year that can be exactly divided by 4 (such as 2012, 2016, etc)
not	except if it can be exactly divided by 100, then it isn't (such as 2100, 2200, etc)
yes	except if it can be exactly divided by 400, then it is (such as 2000, 2400)